requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.2
django-admin-sortable2==2.3.1
django-cleanup==9.0.0
django-environ==0.13.0
django-mysql==4.19.0
django-redis==6.0.0
django-tinymce==5.0.0
easy-thumbnails==2.10.1
feedparser==6.0.12
gunicorn==25.3.0
idna==3.11
mysqlclient==2.2.8
packaging==26.0
pillow==12.2.0
redis==7.2.0
requests==2.33.1
sgmllib3k==1.0.0
slugify==0.0.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
